buf: Revamp buffer handling
 - All constructors take values first now for simpler handling
 - Allow to read from invalid regions (like Lua string.sub()) and
   return empty string if applicable
 - better tests
 - fix docs
 - fix equality tester

+++ research/size.c
@@ -9,9 +9,9 @@
 int main()
 {
 
-	printf( "size_t:      %lu\n", sizeof( size_t ) );
-	printf( "int:         %lu\n", sizeof( int ) );
-	printf( "MAXSIZE:     %lu\n", MAXSIZE );
-	printf( "MAX_SIZET:   %lu\n", MAX_SIZET );
+	printf( "size_t:      %lu\n", (unsigned long) sizeof( size_t ) );
+	printf( "int:         %lu\n", (unsigned long) sizeof( int ) );
+	printf( "MAXSIZE:     %lu\n", (unsigned long) MAXSIZE );
+	printf( "MAX_SIZET:   %lu\n", (unsigned long) MAX_SIZET );
 	return 0;
 }
